Скачивает файлы

+++ EventHandlers/EventBus.cs
@@ -2,7 +2,7 @@
 
 public static class EventBus
 {
-    public static event Action<int> DownloadProgressChanged;
+    public static event Action<double> DownloadProgressChanged;
     public static event Action? UpdateAvailable;
     public static event Action? DownloadStarted;
     public static event Action? DownloadCompleted;
@@ -11,7 +11,7 @@ public static class EventBus
     public static void OnUpdateAvailable() =>
         UpdateAvailable?.Invoke();
 
-    public static void NotifyDownloadProgressChanged(int progress)=>    
+    public static void NotifyDownloadProgressChanged(double progress)=>    
         DownloadProgressChanged?.Invoke(progress);
